+++ src/router/index.tsx
@@ -1,26 +1,26 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import { AnimatePresence } from 'framer-motion';
-import Splash from '../../pages/Splash';
-import ErrorFound from '../notice/NotFound';
+import Splash from '../pages/Splash';
+import ErrorFound from '../components/notice/NotFound';
 import PrivateRoute from './PrivateRoute';
 import UserLimitRoute from './UserLimitRoute';
 import { PRIVATE_ROUTE_PATH, AUTH_PATH } from './routePath';
-import useUserState from '../../hooks/useUserState';
+import useUserState from '../hooks/useUserState';
 
-const Detail = React.lazy(() => import('../../pages/Detail'));
-const Post = React.lazy(() => import('../../pages/Post'));
-const PostingViewer = React.lazy(() => import('../../pages/PostingViewer'));
-const CreateRoom = React.lazy(() => import('../../pages/CreateRoom'));
-const ChatRoom = React.lazy(() => import('../../pages/ChatRoom'));
-const Signup = React.lazy(() => import('../../pages/Signup'));
-const Login = React.lazy(() => import('../../pages/Login'));
-const KakaoLogin = React.lazy(() => import('../KakaoLogin'));
-const Mypage = React.lazy(() => import('../../pages/Mypage'));
-const RoomViewer = React.lazy(() => import('../../pages/RoomViewer'));
-const MyInfoManage = React.lazy(() => import('../../pages/MyInfoManage'));
-const MyPwManage = React.lazy(() => import('../../pages/MyPwManage'));
-const CommentInPost = React.lazy(() => import('../../pages/CommentInPost'));
+const Detail = React.lazy(() => import('../pages/Detail'));
+const Post = React.lazy(() => import('../pages/Post'));
+const PostingViewer = React.lazy(() => import('../pages/PostingViewer'));
+const CreateRoom = React.lazy(() => import('../pages/CreateRoom'));
+const ChatRoom = React.lazy(() => import('../pages/ChatRoom'));
+const Signup = React.lazy(() => import('../pages/Signup'));
+const Login = React.lazy(() => import('../pages/Login'));
+const KakaoLogin = React.lazy(() => import('../components/kakao/KakaoLogin'));
+const Mypage = React.lazy(() => import('../pages/Mypage'));
+const RoomViewer = React.lazy(() => import('../pages/RoomViewer'));
+const MyInfoManage = React.lazy(() => import('../pages/MyInfoManage'));
+const MyPwManage = React.lazy(() => import('../pages/MyPwManage'));
+const CommentInPost = React.lazy(() => import('../pages/CommentInPost'));
 
 const privateRoutePath = [
 	{ id: 1, path: PRIVATE_ROUTE_PATH.COMMENT_IN_POST, element: <CommentInPost /> },
